requirements.txt
asgiref==3.4.1
autopep8==1.5.7
click==8.0.1
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
passlib==1.7.4
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pydantic==1.8.2
python-dotenv==0.19.0
SQLAlchemy==1.4.23
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
